+++ src/mut_bump_vec.rs
@@ -1282,6 +1282,48 @@ where
         }
     }
 
+    /// Returns a vector of the same size as `self`, with function `f` applied to each element in order.
+    ///
+    /// This function only compiles with `U`s with an alignment and size less or equal to `T`'s.
+    ///
+    /// # Examples
+    /// Mapping to a type with an equal alignment and size:
+    /// ```
+    /// # use bump_scope::{ Bump, MutBumpVec };
+    /// # use core::num::NonZero;
+    /// # let mut bump: Bump = Bump::new();
+    /// let a = MutBumpVec::from_iter_in([0, 1, 2], &mut bump);
+    /// let b = a.map_in_place(NonZero::new);
+    /// assert_eq!(format!("{b:?}"), "[None, Some(1), Some(2)]");
+    /// ```
+    ///
+    /// Mapping to a type with a smaller alignment and size:
+    /// ```
+    /// # use bump_scope::{ Bump, MutBumpVec };
+    /// # let mut bump: Bump = Bump::new();
+    /// let a: MutBumpVec<u32> = MutBumpVec::from_iter_in([0, 1, 2], &mut bump);
+    /// let a_capacity = a.capacity();
+    ///
+    /// let b: MutBumpVec<u16> = a.map_in_place(|i| i as u16);
+    /// assert_eq!(b.capacity(), a_capacity * 2);
+    ///
+    /// assert_eq!(b, [0, 1, 2]);
+    /// ```
+    ///
+    /// Mapping to a type with a higher alignment or size won't compile:
+    /// ```compile_fail,E0080
+    /// # use bump_scope::{ Bump, MutBumpVec };
+    /// # let mut bump: Bump = Bump::new();
+    /// let a: MutBumpVec<u16> = MutBumpVec::from_iter_in([0, 1, 2], &mut bump);
+    /// let b: MutBumpVec<u32> = a.map_in_place(|i| i as u32);
+    /// # _ = b;
+    /// ```
+    pub fn map_in_place<U>(self, f: impl FnMut(T) -> U) -> MutBumpVec<'b, 'a, U, A, MIN_ALIGN, UP, GUARANTEED_ALLOCATED> {
+        let MutBumpVec { fixed, bump } = self;
+        let fixed = fixed.map_in_place(f);
+        MutBumpVec { fixed, bump }
+    }
+
     /// Extend the vector by `n` clones of value.
     fn extend_with<B: ErrorBehavior>(&mut self, n: usize, value: T) -> Result<(), B>
     where