+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - **fixed:** divide by zero panic when calling `map_in_place` with a ZST
+- **added:** allow `map_in_place` to ZST's regardless of alignment
 
 ## 0.10.10 (2024-11-10)
 - **added:** default generic parameters to `BumpPoolGuard`